# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JffEhr.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Prescriptions": {
      "get": {
        "tags": [
          "Prescriptions"
        ],
        "parameters": [
          {
            "name": "registrationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prescriptions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrescriptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrescriptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrescriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Prescriptions/{id}": {
      "get": {
        "tags": [
          "Prescriptions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Prescriptions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePrescriptionRequest": {
        "required": [
          "dose",
          "frequency",
          "medicationName",
          "registrationId",
          "scheduledTimes",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "medicationName": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "dose": {
            "maxLength": 60,
            "minLength": 1,
            "type": "string"
          },
          "route": {
            "maxLength": 40,
            "type": "string",
            "nullable": true
          },
          "frequency": {
            "maxLength": 60,
            "minLength": 1,
            "type": "string"
          },
          "scheduledTimes": {
            "minLength": 1,
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PrescriptionDto": {
        "type": "object",
        "properties": {
          "prescriptionId": {
            "type": "string",
            "format": "uuid"
          },
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "medicationName": {
            "type": "string",
            "nullable": true
          },
          "dose": {
            "type": "string",
            "nullable": true
          },
          "route": {
            "type": "string",
            "nullable": true
          },
          "frequency": {
            "type": "string",
            "nullable": true
          },
          "scheduledTimes": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prescribedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "prescribedByName": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}